Improve interactive user shell experience
When using `gle -ui jobname` gle tries to setup the container user with a working $HOME folder and /etc/passwd entries.

If the image contains `git` the working directory for the build will be added to the git safe.directory global setting. Previously this was only set for builds and not the interactive shells

As interactive uses often also want to install packages or alter system files. The gle user now gets granted sudo access (if the image has sudo installed)